refactor(database): ローカル DB のクエストモデル定義をパッケージ移動
